main.py
"""
Main Application Entry Point for RAG PDF Q&A System

This is the main orchestrator that initializes and connects all system components.
Key concepts to learn:
- Application architecture and dependency injection
- Configuration management and environment variables
- Error handling and logging setup
- Component initialization order and dependencies
"""

import os
import sys
import logging
from typing import Optional
from dotenv import load_dotenv

# TODO: Import system components (uncomment when implementing)
# from src.pdf_processor import PDFProcessor
# from src.vector_store import VectorStore
# from src.llm_client import OpenRouterClient
# from src.rag_system import RAGSystem
# from src.gradio_ui import GradioInterface

# Setup logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('rag_system.log'),
        logging.StreamHandler(sys.stdout)
    ]
)

logger = logging.getLogger(__name__)


class RAGApplication:
    """
    Main application class that orchestrates the RAG system.
    
    Learning objectives:
    1. Application architecture and component coordination
    2. Configuration management from environment variables
    3. Dependency injection and initialization patterns
    4. Error handling and system validation
    """
    
    def __init__(self):
        """
        Initialize the RAG application.
        
        TODO: Learn about application initialization patterns:
        - Environment variable management
        - Component dependency resolution
        - Configuration validation
        - Graceful error handling
        """
        # Load environment variables
        load_dotenv()
        
        # TODO: Initialize configuration
        self.config = self._load_configuration()
        
        # TODO: Initialize components
        self.pdf_processor: Optional[PDFProcessor] = None
        self.vector_store: Optional[VectorStore] = None  
        self.llm_client: Optional[OpenRouterClient] = None
        self.rag_system: Optional[RAGSystem] = None
        self.gradio_interface: Optional[GradioInterface] = None
        
        # Application state
        self.is_initialized = False
    
    def _load_configuration(self) -> dict:
        """
        Load and validate application configuration.
        
        TODO: Implement configuration management
        Learning: Environment-based configuration for different deployment environments
        
        Returns:
            Configuration dictionary
            
        TODO Implementation:
        1. Load from environment variables
        2. Validate required settings
        3. Set defaults for optional settings
        4. Handle configuration errors gracefully
        """
        config = {
            # OpenRouter Configuration
            "openrouter_api_key": os.getenv("OPENROUTER_API_KEY", ""),
            "openrouter_base_url": os.getenv("OPENROUTER_BASE_URL", "https://openrouter.ai/api/v1"),
            "default_model": os.getenv("DEFAULT_MODEL", "meta-llama/llama-3.1-8b-instruct:free"),
            
            # Vector Database Configuration
            "chroma_persist_directory": os.getenv("CHROMA_PERSIST_DIRECTORY", "./chroma_db"),
            "embedding_model": os.getenv("EMBEDDING_MODEL", "all-MiniLM-L6-v2"),
            
            # Document Processing Configuration
            "max_pdf_size_mb": int(os.getenv("MAX_PDF_SIZE_MB", "50")),
            "chunk_size": int(os.getenv("CHUNK_SIZE", "1000")),
            "chunk_overlap": int(os.getenv("CHUNK_OVERLAP", "200")),
            "max_search_results": int(os.getenv("MAX_SEARCH_RESULTS", "5")),
            
            # UI Configuration
            "gradio_host": os.getenv("GRADIO_HOST", "127.0.0.1"),
            "gradio_port": int(os.getenv("GRADIO_PORT", "7860")),
            "gradio_share": os.getenv("GRADIO_SHARE", "false").lower() == "true",
        }
        
        # TODO: Validate configuration
        self._validate_configuration(config)
        
        return config
    
    def _validate_configuration(self, config: dict):
        """
        Validate application configuration.
        
        TODO: Implement configuration validation
        Learning: Configuration validation patterns and error handling
        
        Args:
            config: Configuration dictionary to validate
            
        Raises:
            ValueError: If configuration is invalid
        """
        # TODO: Implement validation logic
        errors = []
        
        # Check required settings
        if not config["openrouter_api_key"]:
            errors.append("OPENROUTER_API_KEY is required but not set")
        
        # TODO: Add more validation rules
        # - Check file paths are writable
        # - Validate numeric ranges
        # - Check model availability
        
        if errors:
            raise ValueError(f"Configuration errors: {'; '.join(errors)}")
        
        logger.info("Configuration validation passed")
    
    def initialize_components(self):
        """
        Initialize all system components in correct dependency order.
        
        TODO: Implement component initialization
        Learning: Dependency injection and component lifecycle management
        
        Raises:
            RuntimeError: If component initialization fails
            
        TODO Implementation:
        1. Initialize PDF processor
        2. Initialize vector store
        3. Initialize LLM client
        4. Initialize RAG system
        5. Initialize UI components
        6. Validate all connections
        """
        try:
            logger.info("Initializing RAG system components...")
            
            # TODO: Step 1 - Initialize PDF Processor
            logger.info("Initializing PDF processor...")
            # self.pdf_processor = PDFProcessor(
            #     chunk_size=self.config["chunk_size"],
            #     chunk_overlap=self.config["chunk_overlap"]
            # )
            
            # TODO: Step 2 - Initialize Vector Store
            logger.info("Initializing vector store...")
            # self.vector_store = VectorStore(
            #     persist_directory=self.config["chroma_persist_directory"],
            #     embedding_model=self.config["embedding_model"]
            # )
            
            # TODO: Step 3 - Initialize LLM Client
            logger.info("Initializing LLM client...")
            # self.llm_client = OpenRouterClient(
            #     api_key=self.config["openrouter_api_key"],
            #     base_url=self.config["openrouter_base_url"],
            #     default_model=self.config["default_model"]
            # )
            
            # TODO: Step 4 - Initialize RAG System
            logger.info("Initializing RAG system...")
            # self.rag_system = RAGSystem(
            #     vector_store=self.vector_store,
            #     llm_client=self.llm_client,
            #     max_context_chunks=self.config["max_search_results"]
            # )
            
            # TODO: Step 5 - Initialize Gradio Interface
            logger.info("Initializing web interface...")
            # self.gradio_interface = GradioInterface(
            #     rag_system=self.rag_system
            # )
            
            # TODO: Step 6 - Validate connections
            self._validate_system_health()
            
            self.is_initialized = True
            logger.info("✅ All components initialized successfully!")
            
        except Exception as e:
            logger.error(f"❌ Component initialization failed: {str(e)}")
            raise RuntimeError(f"Failed to initialize RAG system: {str(e)}")
    
    def _validate_system_health(self):
        """
        Validate that all system components are working correctly.
        
        TODO: Implement system health checks
        Learning: Health monitoring and system validation
        
        Raises:
            RuntimeError: If any component is not healthy
        """
        try:
            logger.info("Validating system health...")
            
            # TODO: Validate LLM API connection
            # if not self.llm_client.validate_api_key():
            #     raise RuntimeError("OpenRouter API key validation failed")
            
            # TODO: Validate vector store
            # stats = self.vector_store.get_collection_stats()
            # logger.info(f"Vector store status: {stats}")
            
            # TODO: Test end-to-end pipeline with a simple query
            # test_response = self.rag_system.query("test query")
            # if not test_response:
            #     raise RuntimeError("RAG system test query failed")
            
            logger.info("✅ System health validation passed")
            
        except Exception as e:
            logger.error(f"❌ System health validation failed: {str(e)}")
            raise
    
    def run(self):
        """
        Run the RAG application.
        
        TODO: This is the main entry point for running the application
        Learning: Application lifecycle and execution patterns
        
        TODO Implementation:
        1. Initialize components if not already done
        2. Launch the Gradio interface
        3. Handle graceful shutdown
        """
        try:
            if not self.is_initialized:
                self.initialize_components()
            
            logger.info("🚀 Starting RAG PDF Q&A System...")
            
            # TODO: Launch Gradio interface
            # self.gradio_interface.launch(
            #     share=self.config["gradio_share"],
            #     server_name=self.config["gradio_host"],
            #     server_port=self.config["gradio_port"]
            # )
            
            print("📚 RAG PDF Q&A System is starting...")
            print("TODO: Implement actual component initialization")
            print(f"Configuration loaded: {len(self.config)} settings")
            print("Next steps:")
            print("1. Uncomment component imports")
            print("2. Implement actual initialization logic")
            print("3. Set up environment variables")
            print("4. Test with sample PDFs")
            
        except KeyboardInterrupt:
            logger.info("Application stopped by user")
        except Exception as e:
            logger.error(f"Application error: {str(e)}")
            raise
    
    def shutdown(self):
        """
        Gracefully shutdown the application.
        
        TODO: Implement graceful shutdown
        Learning: Resource cleanup and graceful termination
        """
        logger.info("Shutting down RAG application...")
        
        # TODO: Close database connections
        # TODO: Save any pending state
        # TODO: Clean up temporary files
        
        logger.info("Application shutdown complete")


def main():
    """
    Main entry point for the RAG PDF Q&A system.
    
    TODO: This orchestrates the entire application startup
    Learning: Application entry points and error handling
    """
    try:
        # Create and run the application
        app = RAGApplication()
        app.run()
        
    except Exception as e:
        logger.error(f"Fatal application error: {str(e)}")
        sys.exit(1)


if __name__ == "__main__":
    main()
