requirements.txt
# Core Dependencies
gradio>=4.0.0
chromadb>=0.4.0
openai>=1.0.0  # For OpenRouter compatibility
python-dotenv>=1.0.0

# PDF Processing
pypdf2>=3.0.0
pdfplumber>=0.9.0
langchain>=0.1.0
langchain-community>=0.0.10

# Text Processing & Embeddings
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Optional: Advanced PDF processing
# pymupdf>=1.23.0  # For complex PDF layouts
# unstructured>=0.10.0  # For advanced document parsing
